Fix issue #3, cleanup, new version
Some stupid bugs in the delete fixup. Also made the *Node()
methods private, because they are dangerous. For example
_deleteNode() with a node which is not in the tree might
lead to the  invariant violation where the left node is not
less than the parent node.

But now the Red Black Tree seems to be at least somewhat
usable. All tests pass!

I also split the file redblack.ts into tree.ts and node.ts
and cleaned up the code to make it simpler and stricter.